chore(meshupgrade): extract mesh wide sync calls
